+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+<a name="v0.2.0"></a>
+### v0.2.0 - 2023-07-16
+- value column
+- option to show or not author
 
 <a name="v0.1.0"></a>
 ### v0.1.0 - 2023-07-15